requirements.txt
# Core
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9

# RAG Stack
langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.14
llama-index==0.10.51
llama-index-vector-stores-chroma==0.1.10

# Vector Store
chromadb==0.5.3

# LLM / Embeddings
openai==1.35.7
sentence-transformers==3.0.1

# Document Parsing
pypdf==4.2.0
python-docx==1.1.2

# Evaluation
ragas==0.1.14
datasets==2.20.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.4
tiktoken==0.7.0
